+++ packages/adguard-api/CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.3.0] - 2023-08-10
 ### Changed
 - [BREAKING CHANGE] Assistant script is separated and dynamically injected. You need to provide it in your project.
-- Update `@adguard/tswebextension` to `v0.0.62`
+- Update `@adguard/tswebextension` to `v0.3.9`
+- Update `@adguard/assistant` to `v4.3.70`
 
 
 ## [1.2.2] - 2022-12-02