+++ packages/tswebextension/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version to the end of the file -->
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
+## [0.3.9] - 2023-08-10
+
+### Changed
+- Updated `@adguard/tsurlfilter` to `v2.1.7`.
+
+
 ## [0.3.8] - 2023-08-04
 
 ### Changed